dtg::minimal: turned on UsePAM option when configuring sshd
By turning this on we invoke the pam_motd module when you log in so you
can see the MOTD (and it gets regenerated). I'm using this on nas01 to
show the disk status at the moment.  The configuration leaves
PasswordAuthentication and ChallengeResponseAuthentication turned off so
I don't think we're opening up any new methods by doing this.