Added ~/.ssh/authorized_keys to the list of locations for checking key files
Sometimes I need to create ad-hoc access to a machine for people not on monkeysphere so I need somewhere to put their ssh key.  To do this I needed to switch over to the monkeysphere::sshd class rather than making use of the 'default' helper.  The extra options needed I just copied from the default one.