Written a zpool_status munin plugin.
munin-contrib doesn't seem to have a plugin that monitor the health of
the pool, and throws a wobbly if it appears ill.

This patch will report:
-when pools have devices that are
  online
  degraded
  faulted
  offline
  unavail
  removed
-if any of these, except online, is >0, an error is created.

The exact numbers reported are not quite right. if one disc goes
offline, then the vdev becomes degraded, and the pool does. The
downgrade of the pool is reported twice. However, this is not an issue -
it will just mean that someone will fix it faster.

Moreover there is a missing feature. We should monitor the number of I/O
errors, by reading the READ, WRITE, CKSUM values of `zpool status' and
set an error on this. We can reset the count when a disk is replaced to
get rid of the error.