This patch fixes the bug whereby licking would not work properly, because SSH wants the user to check fingerprints that they are not attached to.
I get around this by creating a wrapper around SSH that adds flags not
to check for fingerprints, and passing all other arguments into SSH. I
then change the GIT_SSH environment variable to point to this script
rather than `builtin ssh`.

This may seem somewhat dangerous. However, it does not appear to be a
security issue. We are only ignoring fingerprints whilst pushing the
puppet config. We *can* be MITMed in this process. But the attacker will
only gain from having our puppet config on his box. Given it is on
github anyway, and has no confidential stuff, we need not care. I just
hope that the attacker enjoys DTG minimal :-).