Run munin-node-configure on each munin node.
Munin will then scan the machine for stuff that it knows how
to monitor, and then monitor it for us.

This is nice as it reduces the amount of config we have to write.
For instance, acr31 had to write repetitive code telling nas01
to report its smart values to munin-gatherer. However, this new
way automagically detects it has smart installed, and says `oh yes
I'll monitor that'.