Make mountd run on a statically defined port on nas04.
By default, mountd will start a server on a random port. You are then
supposed to open all ports on the NFS server to all trusted
clients. This doesn't fit with our use of iptables, opening a
conservative number of ports. Therefore, we adjust this to be static.

This seems to only matter with clients running NFSv3. If you are running
NFSv4, mounting works using the already defined ports. However XenServer
does not yet support NFSv4.

Perhaps we should look into adding this to NAS02.