Fix bugs reporting SSHException: Unable to open channel and ChannelException: Administratively prohibited.
It turns out that fabric opens a new SSH connection for each run()
command, and is asynchronous. Without the patch, we create a tight code
loop that spawns a thread for each VM that runs a snapshot command. The
problem is XenServer has max_logins set to 10. We quickly have 10 open
SSH connections and get these error messages.

The solution to this would be to have one run() command that asks XAPI
to snapshot *all* VMs that we care about. Well actually, XAPI is
documenting as having that feature with xe vm-snapshot, but it doesn't
work due to a bug. Howver, I have patched this and sent a pull request
upstream. Therefore, hopefully this will be fixed in XS-{$NEXT}.

Arguably a nicer short term solution would be a thread pool of 10
threads, but meh. We have done the proper fix, so the hack will do.