Allow DHCP to be bootstrapped.
Without this patch
-A vm is created from a clone, and is booted
-Dom0 attempts to SSH into the vm, but can't because the vm doesn't have
an IP address yet
-The vm can't get an IP address (it's blocked on getting an address from
DHCP)
-Dom0  can't set the vm's hostname, so puppet can't configure the vm as
a dhcp server.

This patch avoids this my stating that puppy8==dhcp. If a vm is booted,
and its hostname is puppy.*, and has a mac address that would be
associated with puppy8, we now set the hostname to be dhcp, and rerun
puppet (which should turn the VM into a dhcp server).