Stage 1 of removing puppet imports.
We now use puppet apply and point it at a directory of pp files, which
is best practice.

Removing the old sites.pp cannot be done yet as we need to ensure all
our machines are using the new post-update hook before removing
it. Remember puppet ensures we are running an up-to-date post update
hook. If we remove sites.pp now, git would delete it from the puppet
repository and then call the old post-update hook, which would call the
broken sites.pp.