Disable ipv6 privacy extensions to be inline with CUDN/JANET rules
MAJ:

By the way, it is lab policy (FWIW) that the use of privacy-enhanced or
other random addresses should be disabled (they are the default in some
operating systems). Machines should normally use EUI-64 addresses based
on the MAC address, and/or explicitly allocated service addresses. The
reason for this is that the CUDN/JANET rules require us to be able to
track an IP address back to a user retrospectively, and without snooping
the logging all neighbour discovery traffic we can't do that if the
addresses are essentially random numbers. We can't enforce this policy,
but it's helpful if machines are appropriately configured where possible.