Revert "Make pre-commit hook execute in the correct directory"
The pre-commit hook runs in the current working directory in which the git commit is run from. Using ../.. might (and indeed for me did) result in it validating all the puppet in all the git repositories I have on my machine not just the dtg-puppet repository. This is very slow and tells you about problems in unrelated repositories.

This reverts commit f11870c35f58b3d5c0cfa812563a59da0b49a7e6.